# server/src/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  #  title: Mentor Success API
  description: API for the Mentor Success application
  contact:
    email: whitney@aiontechnology.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.1.0
servers:
  - url: "https://lj6z99n04d.execute-api.us-west-2.amazonaws.com/mentorsuccess-test-service-deployment"
...
paths:
...
  /api/v1/schools/{schoolId}/students/{studentId}/assessments/{assessmentId}:
    parameters:
      - in: header
        name: Authorization
        schema:
          type: string
      - in: path
        name: schoolId
        schema:
          type: string
          format: uuid
        required: true
        description: The id of the desired school
      - in: path
        name: studentId
        schema:
          type: string
          format: uuid
        required: true
        description: The id of the desired student
      - in: path
        name: assessmentId
        schema:
          type: string
          format: uuid
        required: true
        description: The if the desired assessment workflow

    # OPTIONS for /api/v1/schools/{schoolId}/students/{studentId}/assessments/{assessmentId}
    options:
      tags:
        - school
        - student
      responses:
        200:
          description: Success
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: "string"
            Access-Control-Allow-Methods:
              schema:
                type: "string"
            Access-Control-Allow-Headers:
              schema:
                type: "string"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Methods: "'DELETE,GET,HEAD,OPTIONS,PATCH,POST,PUT'"
              method.response.header.Access-Control-Allow-Headers: "'Content-Type,Authorization,X-Amz-Date,X-Api-Key,X-Amz-Security-Token'"
              method.response.header.Access-Control-Allow-Origin: "'*'"
        requestTemplates:
          application/json: "{\"statusCode\": 200}"
        passthroughBehavior: "when_no_match"
        type: "mock"

    # POST for /api/v1/schools/{schoolId}/students/{studentId}/assessments/{assessmentId}
    post:
      summary: Receive result of assessment
      description: Receive the result of the assessment workflow
      tags:
        - school
        - student
      operationId: receiveAssessmentData
      requestBody:
        description: Assessment data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssessmentDataInputModel'
      responses:
        201:
          description: The post assessments data was received
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/NotAuthenticated'
        403:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/Error'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        uri: "${lb_url}/api/v1/schools/{schoolId}/students/{studentId}/assessments/{assessmentId}"
        responses:
          201:
            statusCode: "201"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          400:
            statusCode: "400"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          401:
            statusCode: "401"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          403:
            statusCode: "403"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          404:
            statusCode: "404"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          default:
            statusCode: "500"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
        requestParameters:
          integration.request.header.Authorization: "method.request.header.Authorization"
          integration.request.header.X-Forwarded-Proto: "'https'"
          integration.request.path.schoolId: "method.request.path.schoolId"
          integration.request.path.studentId: "method.request.path.studentId"
          integration.request.path.assessmentId: "method.request.path.assessmentId"
          integration.request.header.X-Forwarded-Host: "stageVariables.api_url"
        passthroughBehavior: "when_no_match"
        connectionType: "VPC_LINK"
        connectionId: "$${stageVariables.vpc_link_id}"
        httpMethod: "POST"
        type: "http"

  # /api/v1/schools/{schoolId}/students/{studentId}/workflow/assessments/{assessmentId}
  /api/v1/schools/{schoolId}/students/{studentId}/workflow/assessments/{assessmentId}:
    parameters:
      - in: header
        name: Authorization
        schema:
          type: string
      - in: path
        name: schoolId
        schema:
          type: string
          format: uuid
        required: true
        description: The id of the desired school
      - in: path
        name: studentId
        schema:
          type: string
          format: uuid
        required: true
        description: The id of the desired student
      - in: path
        name: assessmentId
        schema:
          type: string
          format: uuid
        required: true
        description: The if the desired assessment workflow

    # OPTIONS for /api/v1/schools/{schoolId}/students/{studentId}/workflow/assessments/{assessmentId}
    options:
      tags:
        - school
        - student
      responses:
        200:
          description: Success
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: "string"
            Access-Control-Allow-Methods:
              schema:
                type: "string"
            Access-Control-Allow-Headers:
              schema:
                type: "string"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Methods: "'DELETE,GET,HEAD,OPTIONS,PATCH,POST,PUT'"
              method.response.header.Access-Control-Allow-Headers: "'Content-Type,Authorization,X-Amz-Date,X-Api-Key,X-Amz-Security-Token'"
              method.response.header.Access-Control-Allow-Origin: "'*'"
        requestTemplates:
          application/json: "{\"statusCode\": 200}"
        passthroughBehavior: "when_no_match"
        type: "mock"

    # GET for /api/v1/schools/{schoolId}/students/{studentId}/workflow/assessments/{assessmentId}
    get:
      summary: Get a assessment workflow for a student
      description: Get an assessment workflow
      tags:
        - school
        - student
      operationId: getAssessmentWorkflow
      responses:
        200:
          description: An assessment workflow
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentWorkflowResultModel'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/NotAuthenticated'
        403:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/Error'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        uri: "${lb_url}/api/v1/schools/{schoolId}/students/{studentId}/workflow/assessments/{assessmentId}"
        responses:
          200:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          400:
            statusCode: "400"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          401:
            statusCode: "401"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          403:
            statusCode: "403"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          404:
            statusCode: "404"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          default:
            statusCode: "500"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
        requestParameters:
          integration.request.header.Authorization: "method.request.header.Authorization"
          integration.request.header.X-Forwarded-Proto: "'https'"
          integration.request.path.schoolId: "method.request.path.schoolId"
          integration.request.path.studentId: "method.request.path.studentId"
          integration.request.path.assessmentId: "method.request.path.assessmentId"
          integration.request.header.X-Forwarded-Host: "stageVariables.api_url"
        passthroughBehavior: "when_no_match"
        connectionType: "VPC_LINK"
        connectionId: "$${stageVariables.vpc_link_id}"
        httpMethod: "GET"
        type: "http"

  # /api/v1/schools/{schoolId}/teachers --------------------------------------------
...
  /api/v1/schools/{schoolId}/workflow/assessments/{assessmentId}:
    parameters:
      - in: header
        name: Authorization
        schema:
          type: string
      - in: path
        name: schoolId
        schema:
          type: string
          format: uuid
        required: true
        description: The id of the desired school
      - in: path
        name: assessmentId
        schema:
          type: string
          format: uuid
        required: true
        description: The id of the desired assessment workflow

    # OPTIONS for /api/v1/schools/{schoolId}/workflow/assessments/{assessmentId}
    options:
      tags:
        - school
      responses:
        200:
          description: Success
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: "string"
            Access-Control-Allow-Methods:
              schema:
                type: "string"
            Access-Control-Allow-Headers:
              schema:
                type: "string"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Methods: "'DELETE,GET,HEAD,OPTIONS,PATCH,POST,PUT'"
              method.response.header.Access-Control-Allow-Headers: "'Content-Type,Authorization,X-Amz-Date,X-Api-Key,X-Amz-Security-Token'"
              method.response.header.Access-Control-Allow-Origin: "'*'"
        requestTemplates:
          application/json: "{\"statusCode\": 200}"
        passthroughBehavior: "when_no_match"
        type: "mock"

    # GET for /api/v1/schools/{schoolId}/workflow/assessments/{assessmentId}
    get:
      summary: Get the assessment workflow
      description: Get the assessment workflow with the given id
      tags:
        - school
      operationId: getAssessmentWorkflow
      responses:
        200:
          description: Returned the workflow
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostAssessmentOutputModel'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/NotAuthenticated'
        403:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/Error'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        uri: "${lb_url}/api/v1/schools/{schoolId}/workflow/assessments/{assessmentId}"
        responses:
          200:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          400:
            statusCode: "400"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          401:
            statusCode: "401"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          403:
            statusCode: "403"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          404:
            statusCode: "404"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          default:
            statusCode: "500"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
        requestParameters:
          integration.request.header.Authorization: "method.request.header.Authorization"
          integration.request.path.schoolId: "method.request.path.schoolId"
          integration.request.path.assessmentId: "method.request.path.assessmentId"
          integration.request.header.X-Forwarded-Proto: "'https'"
          integration.request.header.X-Forwarded-Host: "stageVariables.api_url"
        passthroughBehavior: "when_no_match"
        connectionType: "VPC_LINK"
        connectionId: "$${stageVariables.vpc_link_id}"
        httpMethod: "GET"
        type: "http"

  # /api/v1/tags -------------------------------------------------------------------
...
components:
  responses:
    BadRequest:
      description: There was an error while creating the person
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestModel'
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
    Error:
      description: Encountered a server error
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
    NotAuthenticated:
      description: Access token is missing or invalid
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
    NotAuthorized:
      description: User does not have permissions for operation
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
    NotFound:
      description: Requested resource could not be found
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
...
  schemas:
...
    AssessmentDataInputModel:
      type: object
      properties:
        question1:
          type: number
        question2:
          type: number
        question3:
          type: number
        question4:
          type: number
        question5:
          type: number
        question6:
          type: number
        question7:
          type: number
        question8:
          type: number
        question9:
          type: number
        question10:
          type: number
        question11:
          type: number
        question12:
          type: number
        question13:
          type: number
        question14:
          type: number
        question15:
          type: number
        question16:
          type: number
        question17:
          type: number
        question18:
          type: number
        question19:
          type: number
        question20:
          type: number
        question21:
          type: number
        question22:
          type: number
        question23:
          type: number
        question24:
          type: number
        question25:
          type: number
        question26:
          type: number
        question27:
          type: number
        question28:
          type: number
        question29:
          type: number
        question30:
          type: number
        question31:
          type: number
        question32:
          type: number
        question33:
          type: number
        question34:
          type: number
        question35:
          type: number

    # AssessmentWorkflowResultModel ------------------------------------------------
    AssessmentWorkflowResultModel:
      required:
        - id
        - studentName
      properties:
        id:
          type: string
          format: uuid
          example: 734aa3e8-52e4-488b-8922-09d8eb803ad8
        studentName:
          type: string
          maxLength: 50
          example: Sally
      _links:
        type: object
        allOf:
          - $ref: '#/components/schemas/SelfLink'

    # BadRequestModel --------------------------------------------------------------
    BadRequestModel:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: string
        error:
          type: object
        message:
          type: string
        path:
          type: string

    # BehaviorModelCollection ------------------------------------------------------
...
    Empty:
      type: object

    # GameModel --------------------------------------------------------------------
...
    PostAssessmentOutputModel:
      type: object
      required:
        - id
        - studentName

    # ProgramAdminModel ------------------------------------------------------------
...
    SelfLink:
      type: object
      properties:
        self:
          allOf:
            - $ref: '#/components/schemas/UrlArray'

    # StudentModel -----------------------------------------------------------------
...
    UrlArray:
      type: array
      items:
        type: object
        properties:
          href:
            type: string
            format: url